messing around with the size of the bitmatrix
Are there any benefits to having a super large bitmatrix?
Considering hardcoding a size limit to the QR code to make the performance
hit lower for larger QR codes.

This decouples the size of the QR code away from the size of the screen.
I will need to set the size of the QR code to the screen size instead.

I don't feel the performance hit at 500... I wonder what's the largest
number before I hit the too-laggy threshold.