Adjust image / textbox screen estate ratio + flatten if statements
The image should have more screen size than the text box, as the main
point of the app is the QR code.

getStringFromIntent now has less nested code and is easier to track.

The previously mentioned bug where the text type hint doesn't move has
been fixed. In addition to the change, text hint for the textbox has
been added back to make the lint happy.